\name{ICEinfer-internal}
\alias{ICEinfer-internal}
\alias{ICEangle}
\alias{ICEd2m}
\alias{ICErunif}
\alias{plot.ICEalice}
\alias{plot.ICEscale}
\alias{plot.ICEwedge}
\alias{print.ICEalice}
\alias{print.ICEcolor}
\alias{print.ICEepmap}
\alias{print.ICEscale}
\alias{print.ICEwedge}
\title{Internal ICEinfer Functions.}
\description{Internal method functions for the ICEinfer package.  The plot() and
 print() functions also accept optional \dots arguments documented elsewhere.}
\author{Bob Obenchain <softrx@iquest.net>}
\keyword{methods}
\keyword{internal}